Add `StringArrayCopy.h` for cross-platform utilities; improve WebView teardown safety and modularize `WindowCore`.
- Introduced `StringArrayCopy.h` to handle string array operations across platforms.
- Enhanced WebView teardown process by revoking event tokens and cleaning up references.
- Unified `WindowCore` with modularized initialization and diagnostics improvements.
- Updated CMakeLists to include new utility headers.
