+++ src/InfiniFrame.NativeBridge/Native/Platform/Windows/Core/UiDispatcher.Win32.cpp
@@ -7,31 +7,6 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
-void InfiniFrameWindow::WaitForExit() {
-    auto* impl = m_impl.get();
-    ApplyPendingOwnerWindow(impl, L"wait_for_exit");
-
-    messageLoopRootWindowHandle = impl->_hWnd;
-    TraceTeardown(L"WaitForExit start instance=%p hwnd=%p", this, impl->_hWnd);
-
-    MSG msg = {};
-    while (true) {
-        const int getMessageResult = GetMessage(&msg, nullptr, 0, 0);
-        if (getMessageResult == -1) {
-            TraceTeardown(L"WaitForExit GetMessage failed err=%lu", GetLastError());
-            break;
-        }
-        if (getMessageResult == 0)
-            break;
-
-        TranslateMessage(&msg);
-        DispatchMessage(&msg);
-    }
-
-    messageLoopRootWindowHandle = nullptr;
-    TraceTeardown(L"WaitForExit end instance=%p hwnd=%p", this, impl->_hWnd);
-}
-
 void InfiniFrameWindow::Invoke(ACTION callback) {
     if (!callback) {
         return;
